Task3/arkpz-pzpi-22-2-iuklenchuk-roman-task3/backend/requirements.txt
alembic==1.14.0
anyio==4.6.2.post1
asyncpg==0.30.0
fastapi==0.115.5
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic_core==2.27.1
python-dotenv==1.0.1
SQLAlchemy==2.0.36
uvicorn==0.32.1
loguru==0.7.2
bcrypt==4.2.1
PyJWT==2.10.1
python-multipart==0.0.19
celery==5.4.0
redis==5.2.1
flower==2.0.1
pydantic[email]
pandas==2.2.3
Faker==33.1.0